Fix bugs introduced by update

+++ Mod Files/CSharp/Shared/Missions/DistressEscortMission.cs
@@ -66,7 +66,7 @@ namespace MoreLevelContent.Missions
             if (descriptionWithoutReward != null) { description = descriptionWithoutReward.Replace("[reward]", rewardText); }
         }
 
-        public override int GetBaseReward(Submarine sub)
+        public override float GetBaseReward(Submarine sub)
         {
             if (sub != missionSub)
             {
